requirements.txt
Django>=3.0,<4.0
psycopg2-binary==2.9.3
djangorestframework==3.13.1
redis==4.3.3
celery==5.2.7
drf-yasg==1.20
django-debug-toolbar==3.4.0
django-extensions==3.1.5
Werkzeug==2.1.2
